organize(convert): split code out into more files
